// frontend/src/store/blog/store/action-types.ts
export enum BlogActionTypes {
    GET_BLOG_SUCCESS = 'GET_BLOG_SUCCESS',
    GET_BLOG_FAIL = 'GET_BLOG_FAIL',
    GET_BLOG_START = 'GET_BLOG_START',

    GET_BLOG_DETAIL_SUCCESS = 'GET_BLOG_DETAIL_SUCCESS',
    GET_BLOG_DETAIL_FAIL = 'GET_BLOG_DETAIL_FAIL',
    GET_BLOG_DETAIL_START = 'GET_BLOG_DETAIL_START',

    GET_BLOG_BY_ID_SUCCESS = 'GET_BLOG_BY_ID_SUCCESS',
    GET_BLOG_BY_ID_FAIL = 'GET_BLOG_BY_ID_FAIL',
    GET_BLOG_BY_ID_START = 'GET_BLOG_BY_ID_START',
}